requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
flask-migrate
python-docx>=1.1.2
Babel>=2.14.0
gunicorn>=23.0.0
flask-compress>=1.15
psycopg2-binary>=2.9.9
google-api-python-client>=2.149.0
google-auth>=2.35.0
docusign-esign>=4.0.0
pyjwt>=2.8.0
cryptography>=42.0.0

openpyxl
xlwt
xlrd
xlutils
